move ./models under ./wm_suite and switch to relative imports
