+++ src/wm_suite/models/awd_lstm/configs/QRNN_large_config.json
