+++ src/wm_suite/models/rnn/config.json
