+++ src/wm_suite/models/rnn/model2statedict.py
